requirements.txt
pyusb==1.2.1
pyelftools